DevFest25-WS/P8-ParallelAgents/pyproject.toml
[project]
name = "p8-parallelagents"
version = "0.1.0"
description = "Part 8: Parallel Power with ParallelAgent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-generativeai",
    "python-dotenv",
]

